Create A 50%/100% Done Message for Techs/Civics
NotificationPanel.lua - The beginning of additions to use the
notification system as opposed to the status message system. This can be
improved upon and eventually replace the status messages if desired.
StatusMessagePanel.lua - Creating a LuaEvent to be used in TechTree and
CivicTree to call in order to create a status message. Not fully
implemented is changing the message color depending on the source.
CivicsTree.lua/TechTree.lua - Checks at the beginning of each turn to
see if the current tech/civic has reached 50% or has completed. If so,
create a status message.