Cultural overlay can now be toggled by info button
Ultimately I decided to keep the button from Ace's mod and ditch the base
fading text, since I liked the constant visibility and dedicated tooltip.
Since I was keeping the button anyways I decided to bind it to toggling
the cultural overlay, since it could get somewhat annoying! I plan to add
a setting in the future that changes the default toggle state of the
cultural overlay