Fixes a nasty vanilla bug
The function deal:IsValid() is not trustworthy. It does not always return the expected value, and is most noticeable with some deals for which the function returns False even if the deal is clearly valid and existent. This would manifest as a luxury resource in the Resources tab of the Report Screen identified as coming from "Great People" when in fact the source is another civ. Once the test condition IF deal:IsValid() is removed from the iterator, the resource is correctly identified from the deal item. You can test it with the original file and then with this correction to see the difference.