Added a highly customizable tool for notifications
This method is called the paradoxbar. So far only two test notification
types exist "debug" and "debug2". Keep in mind that this tool is very
flexible and quite capable of both one-off notification types as well as
template-based notifications. Custom behavior should be relatively
simple to implement and each notification falls into a category with
others that fall in the same grouping