Summer 2017 - Unit paths show on hover fix
As soon as devs added queued paths for selected units and because of trade layer paths became thicker I removed a part of CQUI code showing paths on selection and I left showing paths on hover only. Now to show unit paths on hover we use the same layers and the same function as vanilla does when a unit is selected.