Summer 2017 - Unit paths show on hover fix 2
I have no idea why but an issue appeared in Summer-2017 patch when function UnitFlag.SetInteractivity() didn't update unitID for CQUI_ShowPath() function after a unit was upgraded. In that case we couldn't see paths when hover upgraded units. To fix this we add a table with unitIDs and take values from there when units are upgraded.
Also added some tweaks to a previous commit.