Summer 2017 - Unit paths show on hover fix
As soon as devs added queued paths for selected units and because of trade layer paths became thicker I removed a part of CQUI code showing paths on selection and I left showing paths on hover only. Now to show and hide unit paths on hover we use the same layers and the same functions as vanilla does when a unit is selected/deselected.
I have no idea why but an issue appeared in Summer-2017 patch when function UnitFlag.SetInteractivity() didn't update unitID for CQUI_ShowPath() function after a unit was upgraded. In that case we couldn't see paths when hover upgraded units. To fix this we add a table with unitIDs and take values from there when units are upgraded.